feat(Swap): add facade page to auto switch the user to trident when enabled
