+++ src/components/Web3Network/index.tsx
@@ -1,4 +1,3 @@
-import { ChainId } from '@sushiswap/core-sdk'
 import { NETWORK_ICON } from 'app/config/networks'
 import { switchToNetwork } from 'app/functions/network'
 import useIsWindowVisible from 'app/hooks/useIsWindowVisible'
@@ -88,13 +87,7 @@ function Web3Network(): JSX.Element | null {
   // set chainId on initial load if not present
   useEffect(() => {
     if (chainId && !queryChainId) {
-      console.log({ router })
-      if (chainId === ChainId.MATIC && router.route === '/legacy/swap') {
-        router.replace({ pathname: '/trident/swap', query: { ...router.query, chainId } })
-      } else {
-        console.debug('Setting chain id on initial load because not present')
-        router.replace({ pathname: window.location.pathname, query: { ...router.query, chainId } })
-      }
+      router.replace({ pathname: window.location.pathname, query: { ...router.query, chainId } })
     }
   }, [chainId, queryChainId, router])
